+++ ddxApplicationSettings_by_sKeyword/ddxMicrosoftEdgeSettings_by_sKeyword.py
@@ -2,7 +2,7 @@ import os;
 from dxConfig import dxConfig;
 from fsFirstExistingFile import fsFirstExistingFile;
 from mColors import *;
-from mWindowsAPI import oWindowsVersion;
+from mWindowsAPI import oSystemInfo;
 from oConsole import oConsole;
 import mFileSystem;
 
@@ -17,20 +17,28 @@ dxConfigSettings = {
 sEdgeRecoveryPath = mFileSystem.fsPath(os.getenv("LocalAppData"), \
     "Packages", "Microsoft.MicrosoftEdge_8wekyb3d8bbwe", "AC", "MicrosoftEdge", "User", "Default", "Recovery", "Active");
 
+def fEdgeSetup(bFirstRun):
+  if bFirstRun:
+    if oSystemInfo.uOSBuild < 15063:
+      oConsole.fPrint(ERROR, "Debugging Microsoft Edge directly using BugId is only supported on Windows");
+      oConsole.fPrint(ERROR, "builds ", ERROR_INFO, "15063", ERROR, " and higher, and you are running build ", \
+          ERROR_INFO, oSystemInfo.sOSBuild, ERROR, ".");
+      oConsole.fPrint();
+      oConsole.fPrint("You could try using the ", INFO, "EdgeBugId.cmd", NORMAL, " script that comes with EdgeDbg,");
+      oConsole.fPrint("which you can download from ", INFO, "https://github.com/SkyLined/EdgeDbg", NORMAL, ".");
+      oConsole.fPrint("It can be used to debug Edge in BugId on Windows versions before 10.0.15063.");
+      os._exit(4);
+  # Delete the recovery path both the application runs, so as to start with a clean state, and not to keep state
+  # between different runs of the application.
+  fDeleteRecovery();
+
 def fEdgeCleanup():
+  fDeleteRecovery();
+
+def fDeleteRecovery():
+  # Delete the recovery path to clean up after the application ran.
   if mFileSystem.fbIsFolder(sEdgeRecoveryPath):
     mFileSystem.fbDeleteChildrenFromFolder(sEdgeRecoveryPath);
-
-def fEdgeVersionCheck():
-  if oWindowsVersion.uCurrentBuild < 15063:
-    oConsole.fPrint(ERROR, "Debugging Microsoft Edge directly using BugId is only supported on Windows");
-    oConsole.fPrint(ERROR, "builds ", ERROR_INFO, "15063", ERROR, " and higher, and you are running build ", \
-        ERROR_INFO, oWindowsVersion.sCurrentBuild, ERROR, ".");
-    oConsole.fPrint();
-    oConsole.fPrint("You could try using the ", INFO, "EdgeBugId.cmd", NORMAL, " script that comes with EdgeDbg,");
-    oConsole.fPrint("which you can download from ", INFO, "https://github.com/SkyLined/EdgeDbg", NORMAL, ".");
-    oConsole.fPrint("It can be used to debug Edge in BugId on Windows versions before 10.0.15063.");
-    os._exit(4);
 
 def fasGetEdgeOptionalArguments(bForHelp = False):
   return bForHelp and ["<dxConfig.sDefaultBrowserTestURL>"] or [dxConfig["sDefaultBrowserTestURL"]];
@@ -38,13 +46,13 @@ def fasGetEdgeOptionalArguments(bForHelp = False):
 
 ddxMicrosoftEdgeSettings_by_sKeyword = {
   "edge": {
-    "fCheckApplication": fEdgeVersionCheck,
     "dxUWPApplication": {
       "sPackageName": "Microsoft.MicrosoftEdge",
       "sId": "MicrosoftEdge",
     },
     "asApplicationAttachToProcessesForExecutableNames": ["browser_broker.exe"],
     "fasGetOptionalArguments": fasGetEdgeOptionalArguments,
+    "fSetup": fEdgeSetup,
     "fCleanup": fEdgeCleanup,
     "dxConfigSettings": dxConfigSettings,
   },